Address Codex review: keep weekday names dependency-free, type PEM validators
- Extract DAY_NAMES/VALID_DAY_NAMES into a new dependency-free
  shared/day-names.ts. Reusing VALID_DAY_NAMES via dates.ts pulled the
  self-contained catalog-transfer schema into the timezone/settings
  module graph, regressing the cold-start invariant that module's own
  comment documents. Every consumer (including dates.ts itself) now
  imports the constants from day-names.ts directly.
- Annotate isValidPemCertificate/isValidPemPrivateKey with an explicit
  PemValidator return type, per the project's rule that exported
  functions carry an explicit return type rather than relying on
  inference.
