+++ src/features/admin/settings-listing-defaults.ts
@@ -11,7 +11,7 @@ import { t } from "#i18n";
 import { settingsHandler } from "#routes/admin/settings-helpers.ts";
 import { ownerPage } from "#routes/auth.ts";
 import type { TypedRouteHandler } from "#routes/router.ts";
-import { VALID_DAY_NAMES } from "#shared/dates.ts";
+import { VALID_DAY_NAMES } from "#shared/day-names.ts";
 import { invalidateListingsCache } from "#shared/db/listings/records.ts";
 import { settings } from "#shared/db/settings.ts";
 import { isDemoMode } from "#shared/demo/mode.ts";
