+++ test/lib/server-parents-gate/helpers.ts
@@ -12,7 +12,7 @@
 
 import { expect } from "@std/expect";
 import { it as test } from "@std/testing/bdd";
-import { DAY_NAMES } from "#shared/dates.ts";
+import { DAY_NAMES } from "#shared/day-names.ts";
 import { listingChildren } from "#shared/db/listing-parents.ts";
 import type { Listing } from "#shared/types.ts";
 import { bookAttendee } from "#test-utils/db-helpers/attendee-payments.ts";
